Move to editable state
